extract express-validator rules and validate function
